export auto_smooth fix
* Removed remnants of "use_auto_smooth", "auto_smooth_angle" and "calc_normals_split" removed in Blender 4.1.

Btw. As they wrote:
"The Mesh auto_smooth property has been replaced by a modifier node group asset."
I completly don't know if due to this and other changes with Mesh, it should be removed as in this fix, or done in other way.
Someone with more knowledge should look at this.